Additional region support [BGBUILD-120]
This adds region specific constraints and endpoints for s3 bucket
creation, and sets a region specific default url so it does not
have to be specified in the config. For regions to work, aws gem
version >= 2.3.27.

This also fixes an issue where the s3 plugin may attempt to upload
a bundle to a non-existent bucket.